Change the AST representation for blocks.
The previous representation meant we would "eval" the AST,
but I want to traverse the AST and emit interpreter commands
instead. This particular represention will allow us to type
the following into an interpreter session.

{block, foo, noop}
{set, foo, {val, moo, noop}}
{set, foo1, {val, moo1, noop}}
{endblock, noop, noop}

To create the structure:

foo = {
    foo = moo;
    foo1 = moo1;
}

The interpreter will still have to be redesigned to be
more "user friendly" eventually.