Differentiate container nodes from generic lists
Since we don't have any tuple data type, it is safe
to say that container nodes, will have each of its
elements as a tuple.

We can now process OTHER types of lists that don't
follow this pattern as normal data.

This means, empty lists [], will be seen as data values
so the only way to create a container node now is via
erlcfg_node:set/3