Processing list of ASTs 1: First remove 'next' fields in records
I've found that its easier and better to process a list of ASTs, instead
of processing one huge tree. For one, a list of ASTs, allows us to have
a nested list, which automagically, defines a new scope, so we do not
have to bend over backwards to detect scopes, etc.

This first patch, removes the next fields from the record definitions for
the #block{} and #set{} records.

Signed-off-by: Essien Ita Essien <essiene@gmail.com>